pyproject.toml
[project]
name = "feature-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "gigachat>=0.1.39",
    "catboost>=1.2",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "python-dotenv>=1.0",
    "shap>=0.44",
]